chore: apply more light lints to plugins/

+++ plugins/ui/source/ui/text_editor/navigate.cpp
@@ -717,7 +717,7 @@ namespace hex::ui {
                 else
                     i = 0;
             }
-            if ((i32) (direction * i) >= (i32) ((line.size() - 1) * (1 + direction) / 2)) {
+            if ((direction * i) >= (i32) ((line.size() - 1) * (1 + direction) / 2)) {
                 if (lineIndex == (i64)                                         maxLineIndex * (1 + direction) / 2) {
                     if (m_active) {
                         m_active = false;
